backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
celery==5.4.0
redis==5.0.4
google-cloud-storage==2.16.0
google-api-python-client==2.129.0
google-auth==2.29.0
python-multipart==0.0.9
httpx==0.27.0
PyYAML==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.3.0
opencv-python-headless
ultralytics>=8.3.0
